feat(api/applications): doc update api and refactor (BOAS-610)
