+++ apps/api/src/server/applications/application/documents/document.routes.js
@@ -1,0 +1,76 @@
+import { Router as createRouter } from 'express';
+import { asyncHandler } from '../../../middleware/async-handler.js';
+import { trimUnexpectedRequestParameters } from '../../../middleware/trim-unexpected-request-parameters.js';
+import {
+	validateApplicationId,
+	validateFolderIds
+} from '../../application/application.validators.js';
+import { provideDocumentUploadURLs, updateDocuments } from './document.controller.js';
+import {
+	validateDocumentIds,
+	validateDocumentsToUpdateProvided,
+	validateDocumentsToUploadProvided
+} from './document.validators.js';
+
+const router = createRouter();
+
+router.post(
+	'/:id/documents',
+	/*
+        #swagger.tags = ['Applications']
+        #swagger.path = '/applications/{id}/documents'
+        #swagger.description = 'Saves new documents to database and returns location in Blob Storage'
+        #swagger.parameters['id'] = {
+            in: 'path',
+			description: 'Application ID here',
+			required: true,
+			type: 'integer'
+        }
+        #swagger.parameters['body'] = {
+            in: 'body',
+            description: 'Document Details',
+            schema: { $ref: '#/definitions/documentsToSave' }
+        }
+        #swagger.responses[200] = {
+            description: 'Documents that have been saved',
+            schema: { $ref: '#/definitions/documentsAndBlobStorageURLs' }
+        }
+	 */
+	validateApplicationId,
+	validateDocumentsToUploadProvided,
+	validateFolderIds,
+	trimUnexpectedRequestParameters,
+	asyncHandler(provideDocumentUploadURLs)
+);
+
+router.patch(
+	'/:id/documents/update',
+	/*
+        #swagger.tags = ['Applications']
+        #swagger.path = '/applications/{id}/documents/update'
+        #swagger.description = 'Updates the status and/or redaction status on an array of documents'
+        #swagger.parameters['id'] = {
+            in: 'path',
+			description: 'Application ID here',
+			required: true,
+			type: 'integer'
+        }
+        #swagger.parameters['body'] = {
+            in: 'body',
+            description: 'Document Update Details',
+            schema: { $ref: '#/definitions/documentsToUpdateRequestBody' },
+			required: true
+        }
+        #swagger.responses[200] = {
+            description: 'Documents that have been updated',
+            schema: { $ref: '#/definitions/documentsToUpdateRequestBody' }
+        }
+	 */
+	validateApplicationId,
+	validateDocumentsToUpdateProvided,
+	validateDocumentIds,
+	trimUnexpectedRequestParameters,
+	asyncHandler(updateDocuments)
+);
+
+export { router as documentRoutes };
